feat: block index migration script
This script reads from the source dynamo table and writes to the destination table.

It transforms the data to the newer index format, attempts to upgrade the index carPath to the car cid format where possible, and avoids writing to the destination table where an equivalant record already exists.

License: MIT
Signed-off-by: Oli Evans <oli@protocol.ai>