feat: block index migration script (#1)
Provides 2 commands

- clis.js: scan the source dynamo table. for each record transform it to
the new format and see if it exists in the destination table. If not,
write the record to disk as ndjson
- write-cli.js: takes an ndjson file of indexes as input and writes them
in batches to the destination table. Writes any operations to disk as
ndjson for retrying.

_these would be a single command, but I am in the middle of running the
read side of things and didn't want to break it, so the write side lives
in separate files. If we care we can merge them later_

Dynamo table scans let us split up the results into segments. You tell
it how many segments to divide the result set into, and a current
segment, and the scan command will fetch all the results for that
segment only. This is nice, as it makes it easier to divide the work up
across workers.

License: MIT

---------

Signed-off-by: Oli Evans <oli@protocol.ai>
Co-authored-by: Alan Shaw <alan.shaw@protocol.ai>